+++ src/layout/Main.jsx
@@ -1,0 +1,16 @@
+import NavBar from "../components/NavBar"
+import Footer from "../components/Footer"
+
+
+export default function Main({children}) {
+  return (
+    <>
+    <NavBar/>
+    {
+        children
+    }
+    <Footer/>
+    </>
+  )
+}
+
