+++ src/pages/Index.jsx
@@ -1,0 +1,11 @@
+import Welcome from "../components/Welcome"
+import Carousel from "../components/Carousel"
+
+export default function Index() {
+  return (
+    <>
+    <Welcome/>
+    <Carousel/>
+    </>
+  )
+}
