+++ tailwind.config.js
@@ -7,11 +7,8 @@ module.exports = {
   theme: {
   extend: {
     backgroundImage: {
-      'img-nav': 'url("public/imagenes/img-nav.png")'
+      'img-nav': 'url("src/imagenes/img-nav.png")'
     },
-    Image: {
-      'nav-mob': 'url("url("public/imagenes/Nav-mobile.png")'
-    }
   }
   },
   variants: {},
